requirements.txt
bcrypt
cryptography
requests
schedule
python-nmap
